Factor LLVM dependency out of Target.h,cpp files.
