+++ Makefile
@@ -263,6 +263,7 @@ SOURCE_FILES = \
   JITModule.cpp \
   Lerp.cpp \
   LinearSolve.cpp \
+  LLVM_Runtime_Linker.cpp \
   Lower.cpp \
   Memoization.cpp \
   ModulusRemainder.cpp \
@@ -359,6 +360,7 @@ HEADER_FILES = \
   Lambda.h \
   Lerp.h \
   LinearSolve.h \
+  LLVM_Runtime_Linker.h \
   Lower.h \
   MainPage.h \
   Memoization.h \
